iommu: arm-smmu: fix compile warning
Bug 1823745

Reviewed-on: http://git-master/r/1238546
(cherry picked from commit 719dcf9fb6a3ca2daca280193b30c2d1f26d6778)
Reviewed-on: http://git-master/r/1287922
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/iommu/arm-smmu.c
@@ -1450,7 +1450,7 @@ static struct iommu_domain *arm_smmu_domain_alloc(unsigned type)
 
 	smmu_domain->arm_dummy_page = alloc_page(GFP_KERNEL | __GFP_ZERO);
 	if (!smmu_domain->arm_dummy_page) {
-		free_pages(pgd, order);
+		free_pages((ulong)pgd, order);
 		goto out_free_domain;
 	}
